权限控制

+++ app/Http/Controllers/PostController.php
@@ -52,7 +52,9 @@ class PostController extends Controller
             'content' => 'required|string|min:10'
         ]);
 
-        $post = Post::create(request(['title', 'content']));
+        $user_id = \Auth::id();
+        $params = array_merge(request(['title', 'content']), compact('user_id'));
+        $post = Post::create($params);
 
         return redirect('/posts');
     }
@@ -79,6 +81,9 @@ class PostController extends Controller
             'title' => 'required|string|max:100|min:5',
             'content' => 'required|string|min:10'
         ]);
+
+        $this->authorize('update', $post);
+
         $post->title = request('title');
         $post->content = request('content');
         $post->save();
@@ -93,7 +98,7 @@ class PostController extends Controller
      */
     public function delete(Post $post)
     {
-        //TODO:用户权限验证
+        $this->authorize('delete', $post);
 
         $post->delete();
